refactor: split region map shared part into constants and visibility (#3878)
Move fog, reveal-halo, label, and overlay helpers out of the 608-line
region_map_component_shared.dart part so each fragment stays under the
600-line flame subtree target.

Refs #3878
